# docs/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GHCAA.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/financials/fees/applicable": {
      "get": {
        "tags": [
          "Financials"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/FinancialCategory"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/MembershipType"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/financials/record-payment": {
      "post": {
        "tags": [
          "Financials"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "PaidAt",
                  "TransactionId"
                ],
                "type": "object",
                "properties": {
                  "TransactionId": {
                    "maxLength": 200,
                    "type": "string"
                  },
                  "Amount": {
                    "minimum": 0.01,
                    "type": "number",
                    "format": "double"
                  },
                  "PaidAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "FinancialCategory": {
                    "$ref": "#/components/schemas/FinancialCategory"
                  },
                  "PaymentMethod": {
                    "$ref": "#/components/schemas/PaymentMethod"
                  },
                  "Notes": {
                    "maxLength": 500,
                    "type": "string"
                  },
                  "MemberId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "Receipt": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "TransactionId": {
                  "style": "form"
                },
                "Amount": {
                  "style": "form"
                },
                "PaidAt": {
                  "style": "form"
                },
                "FinancialCategory": {
                  "style": "form"
                },
                "PaymentMethod": {
                  "style": "form"
                },
                "Notes": {
                  "style": "form"
                },
                "MemberId": {
                  "style": "form"
                },
                "Receipt": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/financials/update-status/{id}": {
      "patch": {
        "tags": [
          "Financials"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/PaymentStatus"
            }
          },
          {
            "name": "notes",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/financials/receipt/{paymentId}": {
      "get": {
        "tags": [
          "Financials"
        ],
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/financials/dues/generate": {
      "post": {
        "tags": [
          "Financials"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/financials/fees/config": {
      "get": {
        "tags": [
          "Financials"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Financials"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMembershipFeeConfigDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMembershipFeeConfigDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMembershipFeeConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Financials"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMembershipFeeConfigDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMembershipFeeConfigDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMembershipFeeConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/financials/membership-history/{memberId}": {
      "get": {
        "tags": [
          "Financials"
        ],
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/financials/payment/{id}": {
      "delete": {
        "tags": [
          "Financials"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/financials/member/{memberId}/history": {
      "get": {
        "tags": [
          "Financials"
        ],
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/financials/saved-methods": {
      "get": {
        "tags": [
          "Financials"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Financials"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSavedPaymentMethodDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSavedPaymentMethodDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSavedPaymentMethodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/financials/saved-methods/{id}": {
      "delete": {
        "tags": [
          "Financials"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateMembershipFeeConfigDto": {
        "required": [
          "effectiveDate",
          "membershipType"
        ],
        "type": "object",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/FinancialCategory"
          },
          "membershipType": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "amount": {
            "minimum": 0.01,
            "type": "number",
            "format": "double"
          },
          "effectiveDate": {
            "type": "string",
            "format": "date-time"
          },
          "effectiveTo": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "description": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateSavedPaymentMethodDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "method": {
            "type": "string",
            "nullable": true
          },
          "accountNumber": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FinancialCategory": {
        "enum": [
          "MembershipFee",
          "RegistrationFee",
          "Donation",
          "Event",
          "Maintenance",
          "Salary",
          "Utilities",
          "ReunionFee",
          "Sponsorship",
          "Grant",
          "Refund",
          "Other"
        ],
        "type": "string"
      },
...
      "MembershipType": {
        "enum": [
          "Founding",
          "Executive",
          "General",
          "Associate",
          "Honorary",
          "Advisory",
          "Guest"
        ],
        "type": "string"
      },
...
      "PaymentMethod": {
        "enum": [
          "ManualReceipt",
          "BKash",
          "Nagad",
          "Rocket",
          "CreditCard",
          "BankTransfer",
          "CashOnHand"
        ],
        "type": "string"
      },
      "PaymentStatus": {
        "enum": [
          "Pending",
          "Completed",
          "Failed",
          "Refunded"
        ],
        "type": "string"
      },
...
      "UpdateMembershipFeeConfigDto": {
        "required": [
          "effectiveDate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "$ref": "#/components/schemas/FinancialCategory"
          },
          "amount": {
            "minimum": 0.01,
            "type": "number",
            "format": "double"
          },
          "effectiveDate": {
            "type": "string",
            "format": "date-time"
          },
          "effectiveTo": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "description": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}